Update BingoBoard.py - Add editable titles
The left-hand screen title can now be passed as a quoted string
on the command line.

An input field was added on the interactive screen to allow the
right-hand side screen title to be dynamically changed. This
allows the type of game to be displayed on the board (Four Corners, Cover-All, ...) It would also allow a short message to be
displayed. (about 20 characters)